pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
python-dotenv = "^1.0"
mysqlclient = "^2.2"
djangorestframework = "^3.15"
bleach = "^6.1"

[tool.poetry.dev-dependencies]
drf-spectacular = "^0.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
